experiments/temporal-models/smokeynet-adapted/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smokeynet_adapted"]

[project]
name = "smokeynet-adapted"
version = "0.1.0"
description = "Adapted SmokeyNet (YOLO + LSTM + ViT) for temporal smoke detection"
requires-python = ">=3.11"
dependencies = [
    "pyrocore",
    "huggingface-hub>=0.24",
    "lightning>=2.2",
    "matplotlib>=3.8",
    "numpy>=1.26,<2",
    "opencv-python-headless>=4.8",
    "pyyaml>=6.0",
    "tensorboard>=2.16",
    "torch>=2.2",
    "torchvision>=0.17",
    "tqdm>=4.66",
    "ultralytics>=8.3",
]

[tool.uv.sources]
pyrocore = { path = "../../../lib/pyrocore" }

[dependency-groups]
dev = [
    "dvc[s3]>=3.56",
    "jupyterlab>=4.0",
    "nbqa>=1.9",
    "nbstripout>=0.8",
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "PLC0415"]

[tool.ruff.format]
quote-style = "double"
